backend/requirements.txt
coverage==4.5.3
Django==2.2.2
djangorestframework==3.7.7
flake8==3.7.7
flake8-quotes==2.0.1
psycopg2==2.7.3.2
pyflakes==2.1.1
pytz==2017.3
pillow==6.1.0
requests
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
